backend/requirements.txt
Django==3.2.4
psycopg2>=2.9,<3.0
daphne==3.0.2
django-ninja==0.13.2
channels==3.0.3
